\name{cork}
\docType{data}
\alias{cork}

\title{Cork borings}
\description{
  Measurements of the weight of cork borings taken from the north (N), east (E), south (S), and west (W) directions of 28 trees.
  It is of interest to compare the bark thickness (and hence weight) in the four directions.
}
\usage{data(cork)}
\format{
  A data frame with 28 observations on the following 4 variables.
  \describe{
    \item{N}{ north.}
    \item{E}{ east.}
    \item{S}{ south.}
    \item{W}{ west.}
  }
}
\source{
  Mardia, K.V., Kent, J.T., and Bibby, J.M. (1979).
  \emph{Multivariate Analysis}. Academic Press, London.
}
\keyword{datasets}
