\name{examScor}
\docType{data}
\alias{examScor}

\title{Open/Closed book data}
\description{
  Dataset from Mardia, Kent and Bibby on 88 students who took examinations in five subjects. The first two subjects were tested
  with closed book exams and the last three were tested with open book exams.
}
\usage{data(examScor)}
\format{
  A data frame with 88 observations on the following 5 variables.
  \describe{
    \item{mechanics}{ mechanics, closed book exam.}
    \item{vectors}{ vectors, closed book exam.}
    \item{algebra}{ algebra, open book exam.}
    \item{analysis}{ analysis, open book exam.}
    \item{statistics}{ statistics, open book exam.}
  }
}
\source{
  Mardia, K.V., Kent, J.T., and Bibby, J.M. (1979).
  \emph{Multivariate Analysis}. Academic Press, London.
}
\keyword{datasets}
