\name{PFM}
\docType{data}
\alias{PFM}

\title{Returns from the Chilean Pension Funds}
\description{
  Dataset from Osorio and Galea (2015) on the Chilean pension funds. The data corresponds to monthly returns of
  pension funds of four Pension Fund Managers (PFMs), namely Cuprum, Habitat, PlanVital and ProVida corresponding
  to the funds A, B, C, D and E, decreasing in terms of risk, in the period from August 2005 to December 2013.
}
\usage{data(PFM)}
\format{
  A data frame with 101 observations and 21 columns including the \code{date} and variables \code{cuprum}, \code{habitat},
  \code{planvital} and \code{provida}, each variable is related to the funds A, B, C, D, and E.
}
\source{
  Osorio, F., and Galea, M. (2015).
  Statistical inference in multivariate analysis using the t-distribution.
  Unpublished manuscript.
}
\keyword{datasets}
