\name{WindSpeed}
\docType{data}
\alias{WindSpeed}

\title{Wind speed data}
\description{
  This dataset consists of 278 hourly average wind speed in the Pacific North-West of the 
  United States collected at three meteorological towers approximately located on a line 
  and ordered from west to east: Goodnoe Hills (gh), Kennewick (kw), and Vansycle (vs). 
  The data were collected from 25 February to 30 November 2003 recorded at midnight, a 
  time when wind speeds tend to peak.
}
\usage{data(WindSpeed)}
\format{
  A data frame with 278 observations on the following 3 variables.
  \describe{
    \item{gh}{ Goodnoe Hills.}
    \item{kw}{ Kennewick.}
    \item{vs}{ Vansycle.}
  }
}
\source{
  Azzalini, A., Genton, M.G. (2008).
  Robust likelihood methods based on the skew-t and related distributions.
  \emph{International Statistical Review} \bold{76}, 106-129.
}
\keyword{datasets}